api-gateway/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0

strawberry-graphql[asgi]>=0.230.0

sqlalchemy>=2.0.23
pyodbc>=5.3.0
python-dotenv>=1.0.0

grpcio>=1.78.0
grpcio-tools>=1.78.0

Faker>=20.1.0
pydantic-core>=2.14.0
annotated-types>=0.6.0

slowapi>=0.1.9
redis>=5.0.1
pybreaker>=1.4.0

prometheus-client>=0.19.0
python-json-logger>=2.0.7

PyJWT>=2.8.0
bcrypt>=4.1.0
cryptography>=41.0.0

black>=23.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0